# docs/api-specs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AuthKit APIs Documentation",
    "version": "1.0.0",
    "description": "API documentation for AuthKit authentication and authorization."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API Server (development)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "csrfAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "googleOAuth": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth",
            "tokenUrl": "https://oauth2.googleapis.com/token",
            "scopes": {
              "openid": "OpenID Connect",
              "profile": "User profile Information",
              "email": "User Email Information"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": [],
      "csrfAuth": []
    }
  ],
  "paths": {
...
    "/oauth/google": {
      "get": {
        "tags": ["OAuth APIs"],
        "summary": "Initiate Google OAuth",
        "description": "Redirects the user to Google for authentication. In Swagger UI, this can be triggered via the \"Authorize\" button if googleOAuth is configured.",
        "security": [
          {
            "googleOAuth": ["openid", "profile", "email"]
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "redirectUrl",
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Optional redirect URL to return to after login"
          },
          {
            "in": "query",
            "name": "uid",
            "schema": {
              "type": "string"
            },
            "description": "Optional OIDC interaction UID"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to Google Login"
          }
        }
      }
    },
    "/oauth/google/callback": {
      "get": {
        "tags": ["OAuth APIs"],
        "summary": "Google OAuth callback",
        "description": "Handles the callback from Google OAuth authentication.",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "description": "The authorization code returned by Google"
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            },
            "description": "The state parameter for CSRF protection"
          }
        ],
        "responses": {
          "200": {
            "description": "OAuth login successful, redirects to frontend"
          },
          "401": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}